requirements.txt
aiogram==2.14.3
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
black==21.8b0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
idna==3.2
isort==5.9.3
multidict==5.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.3.0
python-dotenv==0.19.0
pytz==2021.1
regex==2021.8.28
tomli==1.2.1
typing-extensions==3.10.0.2
yarl==1.6.3
